sjp-event/sjp-event-processor/src/test/java/uk/gov/moj/cpp/sjp/event/processor/listener/CaseUpdatedListenerTest.java
package uk.gov.moj.cpp.sjp.event.processor.listener;

import static org.mockito.Matchers.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static uk.gov.moj.cpp.sjp.event.processor.listener.CaseUpdatedListener.DEFENDANT_ADDED_PUBLIC_EVENT;
import static uk.gov.moj.cpp.sjp.event.processor.listener.CaseUpdatedListener.DEFENDANT_ADDITION_FAILED_PUBLIC_EVENT;
import static uk.gov.moj.cpp.sjp.event.processor.listener.EventProcessorConstants.CASE_ID;
import static uk.gov.moj.cpp.sjp.event.processor.listener.EventProcessorConstants.DEFENDANT_ID;
import static uk.gov.moj.cpp.sjp.event.processor.listener.EventProcessorConstants.DESCRIPTION;

import uk.gov.justice.services.core.enveloper.Enveloper;
import uk.gov.justice.services.core.sender.Sender;
import uk.gov.justice.services.messaging.JsonEnvelope;
import uk.gov.moj.cpp.sjp.event.processor.listener.CaseUpdatedListener;

import java.util.UUID;
import java.util.function.Function;

import javax.json.JsonObject;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.runners.MockitoJUnitRunner;

@RunWith(MockitoJUnitRunner.class)
public class CaseUpdatedListenerTest {

    @InjectMocks
    private CaseUpdatedListener listener;

    @Mock
    private Sender sender;
    @Mock
    private Enveloper enveloper;
    @Mock
    private JsonEnvelope jsonEnvelope;
    @Mock
    private JsonObject payload;
    @Mock
    private Function<Object, JsonEnvelope> enveloperFunction;
    @Mock
    private JsonEnvelope finalEnvelope;

    @Test
    public void shouldHandleDefendantAddedEventMessage() throws Exception {
        when(jsonEnvelope.payloadAsJsonObject()).thenReturn(payload);
        when(payload.getString(CASE_ID)).thenReturn(UUID.randomUUID().toString());
        when(payload.getString(DEFENDANT_ID)).thenReturn(UUID.randomUUID().toString());

        when(jsonEnvelope.payloadAsJsonObject()).thenReturn(payload);
        when(enveloper.withMetadataFrom(jsonEnvelope, DEFENDANT_ADDED_PUBLIC_EVENT)).thenReturn(enveloperFunction);
        when(enveloperFunction.apply(any(JsonObject.class))).thenReturn(finalEnvelope);

        listener.handleDefendantAddedEvent(jsonEnvelope);
        verify(sender).send(finalEnvelope);
        verify(payload).getString(DEFENDANT_ID);
        verify(payload).getString(CASE_ID);
    }

    @Test
    public void shouldHandleDefendantAdditionFailedEventMessage() throws Exception {
        when(jsonEnvelope.payloadAsJsonObject()).thenReturn(payload);
        when(payload.getString(CASE_ID)).thenReturn(UUID.randomUUID().toString());
        when(payload.getString(DEFENDANT_ID)).thenReturn(UUID.randomUUID().toString());
        when(payload.getString(DESCRIPTION)).thenReturn("some description");

        when(jsonEnvelope.payloadAsJsonObject()).thenReturn(payload);
        when(enveloper.withMetadataFrom(jsonEnvelope, DEFENDANT_ADDITION_FAILED_PUBLIC_EVENT)).thenReturn(enveloperFunction);
        when(enveloperFunction.apply(any(JsonObject.class))).thenReturn(finalEnvelope);

        listener.handleDefendantAdditionFailedEvent(jsonEnvelope);
        verify(sender).send(finalEnvelope);
        verify(payload).getString(CASE_ID);
        verify(payload).getString(DEFENDANT_ID);
        verify(payload).getString(DESCRIPTION);
    }

}